requirements.txt
certifi
charset-normalizer==3.4.2
customtkinter==5.2.2
darkdetect==0.8.0
geopy
packaging==25.0
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
Pillow
